Includes refactoring. Common algorithms should not depend on search_engine_data.hpp and contiguous_internalmem_datafacade.hpp
