+++ include/engine/plugins/match.hpp
@@ -2,12 +2,10 @@
 #define MATCH_HPP
 
 #include "engine/api/match_parameters.hpp"
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
 #include "engine/plugins/plugin_base.hpp"
 #include "engine/routing_algorithms.hpp"
 
-#include "engine/map_matching/bayes_classifier.hpp"
-#include "engine/routing_algorithms/map_matching.hpp"
-#include "engine/routing_algorithms/shortest_path.hpp"
 #include "util/json_util.hpp"
 
 #include <vector>
