+++ include/engine/plugins/trip.hpp
@@ -4,6 +4,7 @@
 #include "engine/plugins/plugin_base.hpp"
 
 #include "engine/api/trip_parameters.hpp"
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
 #include "engine/routing_algorithms.hpp"
 
 #include "util/json_container.hpp"
