+++ include/engine/plugins/viaroute.hpp
@@ -4,7 +4,7 @@
 #include "engine/plugins/plugin_base.hpp"
 
 #include "engine/api/route_api.hpp"
-#include "engine/datafacade/datafacade_base.hpp"
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
 #include "engine/routing_algorithms.hpp"
 #include "engine/search_engine_data.hpp"
 #include "util/json_container.hpp"
