+++ include/engine/routing_algorithms/tile_turns.hpp
@@ -3,7 +3,6 @@
 
 #include "engine/algorithm.hpp"
 #include "engine/datafacade.hpp"
-#include "engine/datafacade/datafacade_base.hpp"
 
 #include "util/coordinate.hpp"
 #include "util/typedefs.hpp"
