+++ src/engine/routing_algorithms/alternative_path_ch.cpp
@@ -1,6 +1,9 @@
 #include "engine/routing_algorithms/alternative_path.hpp"
 #include "engine/routing_algorithms/routing_base_ch.hpp"
 
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
+#include "engine/search_engine_data.hpp"
+
 #include "util/integer_range.hpp"
 
 #include <boost/assert.hpp>
