+++ src/engine/routing_algorithms/alternative_path_mld.cpp
@@ -1,6 +1,9 @@
 #include "engine/routing_algorithms/alternative_path.hpp"
 #include "engine/routing_algorithms/routing_base_mld.hpp"
 
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
+#include "engine/search_engine_data.hpp"
+
 #include "util/static_assert.hpp"
 
 #include <boost/assert.hpp>
