+++ src/engine/routing_algorithms/direct_shortest_path.cpp
@@ -4,6 +4,9 @@
 #include "engine/routing_algorithms/routing_base_ch.hpp"
 #include "engine/routing_algorithms/routing_base_mld.hpp"
 
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
+#include "engine/search_engine_data.hpp"
+
 namespace osrm
 {
 namespace engine
