+++ src/engine/routing_algorithms/many_to_many.cpp
@@ -1,6 +1,9 @@
 #include "engine/routing_algorithms/many_to_many.hpp"
 #include "engine/routing_algorithms/routing_base_ch.hpp"
 
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
+#include "engine/search_engine_data.hpp"
+
 #include <boost/assert.hpp>
 
 #include <limits>
