+++ src/engine/routing_algorithms/map_matching.cpp
@@ -2,6 +2,9 @@
 #include "engine/routing_algorithms/routing_base_ch.hpp"
 #include "engine/routing_algorithms/routing_base_mld.hpp"
 
+#include "engine/datafacade/contiguous_internalmem_datafacade.hpp"
+#include "engine/search_engine_data.hpp"
+
 #include "engine/map_matching/hidden_markov_model.hpp"
 #include "engine/map_matching/matching_confidence.hpp"
 #include "engine/map_matching/sub_matching.hpp"
